+++ src/components/index.js
@@ -2,3 +2,5 @@ export { default as Header } from './Header';
 export { default as Input } from './Input';
 export { default as PublicRoute } from './PublicRoute';
 export { default as PrivateRoute } from './PrivateRoute';
+export { default as UserHome } from './UserHome';
+export { default as CafeHome } from './CafeHome';
